+++ package.json
@@ -7,7 +7,8 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "cli:env": "node src/cli/env",
     "cli:args": "node src/cli/args",
-    "hash:calc": "NODE_ENV=development node src/hash/calcHash"
+    "hash:calc": "NODE_ENV=development node src/hash/calcHash",
+    "module:esm": "node src/modules/cjsToEsm.mjs"
   },
   "repository": {
     "type": "git",
